overwrite in one component Sort
